src/requirements.txt
autopep8==1.5
avro==1.9.1
avro-python3==1.9.1
certifi==2019.11.28
chardet==3.0.4
confluent-kafka==1.3.0
idna==2.8
pycodestyle==2.5.0
PyYAML==5.3
requests==2.22.0
urllib3==1.25.8
